@@ -1,12 +1,12 @@
 import logging
+from collections.abc import Container, Iterator
 from itertools import chain
 from pathlib import Path
-from typing import Iterator, Container
 
 import more_itertools
 
 from tag_space_tools.core.tag_finder import TagFinder
-from tag_space_tools.core.tag_space_entry import Tag
+from tag_space_tools.core.tag_space_entry import Tag, TagSpaceEntry
 
 logger = logging.getLogger(__name__)
 
@@ -19,16 +19,20 @@ def _dirGen(sortTag: list[Tag], tags: list[Tag], destDir: Path) -> Iterator[Path
             yield curPath
 
 
-def sortFiles(sortTag: list[Tag],
-              srcDir: Path | str,
-              dstDir: Path | str,
-              maxFilesInDir=20,
-              recursive=True,
-              excludeDir: Container[Path] = (),
-              tagFinderDst: TagFinder = None,
-              tagFinderSrc: TagFinder = None):
+def sortFiles(
+    sortTag: list[Tag],
+    srcDir: Path | str,
+    dstDir: Path | str,
+    maxFilesInDir=20,
+    *,
+    recursive=True,
+    excludeDir: Container[Path] = (),
+    tagFinderDst: TagFinder | None = None,
+    tagFinderSrc: TagFinder | None = None,
+):
     """
     Move files and their tags from 'srcDir' to 'dstDir'.
+
     Subdirectory name is chosen based on moving file tags and 'sortTag' argument.
 
     :param sortTag: tag list defining order of sort
@@ -41,19 +45,21 @@ def sortFiles(sortTag: list[Tag],
     :param tagFinderDst: cached object for better performance
     :param tagFinderSrc: cached object for better performance
     """
-    dstDir = Path(dstDir)
-    dstDir.mkdir(exist_ok=True, parents=True)
+    dstDirPath = Path(dstDir)
+    dstDirPath.mkdir(exist_ok=True, parents=True)
     if tagFinderDst is None:
-        tagFinderDst = TagFinder(dstDir)
+        tagFinderDst = TagFinder(dstDirPath)
 
     if tagFinderSrc is None:
         tagFinderSrc = TagFinder(srcDir)
 
     for tagEntry in tagFinderSrc.getTagEntryInDir(srcDir, recursive=recursive):
 
-        lastDir = dstDir
-        dirGen: Iterator[Path | None] = _dirGen(sortTag, tagEntry.tags, dstDir)
-        for lastDir, nextDir in more_itertools.windowed(chain(dirGen, (None,)), 2):
+        lastDir = dstDirPath
+        dirGen: Iterator[Path | None] = _dirGen(sortTag, tagEntry.tags, dstDirPath)
+        windowed = more_itertools.windowed(chain(dirGen, (None,)), 2)
+
+        for lastDir, nextDir in windowed:
             if nextDir and nextDir.exists():
                 continue
 
@@ -63,16 +69,32 @@ def sortFiles(sortTag: list[Tag],
             if lastDir in excludeDir:
                 continue
 
-            taggedFilesCount = len(tagFinderDst.getTagEntryInDir(lastDir, recursive=False))
+            taggedFilesCount = len(
+                tagFinderDst.getTagEntryInDir(lastDir, recursive=False)
+            )
             if taggedFilesCount < maxFilesInDir:
                 tagEntry.move(lastDir)
                 break
-            elif taggedFilesCount >= maxFilesInDir:
+
+            if taggedFilesCount >= maxFilesInDir:
                 tagEntry.move(lastDir)
-                sortFiles(sortTag, lastDir, dstDir, maxFilesInDir,
-                          recursive=False, excludeDir=(lastDir,),
-                          tagFinderDst=tagFinderDst, tagFinderSrc=tagFinderSrc)
+                sortFiles(
+                    sortTag,
+                    lastDir,
+                    dstDir,
+                    maxFilesInDir,
+                    recursive=False,
+                    excludeDir=(lastDir,),
+                    tagFinderDst=tagFinderDst,
+                    tagFinderSrc=tagFinderSrc,
+                )
                 break
 
         else:
-            tagEntry.move(lastDir)
+            _moveFile(tagEntry, lastDir)
+
+
+def _moveFile(tagEntry: TagSpaceEntry, lastDir: Path | None):
+    if lastDir is None:
+        raise TypeError
+    tagEntry.move(lastDir)
