@@ -1,8 +1,9 @@
+from collections.abc import Iterable
 from pathlib import Path
-from typing import Iterable
 
 from pyqt_utils.python.typing_const import StrPath
 from pyqt_utils.widgets.tag_filter.nodes import TagFilterNode
+
 from tag_space_tools.core.file_searcher import TagSpaceSearcher
 from tag_space_tools.core.tag_space_entry import TagSpaceEntry
 
@@ -18,17 +19,19 @@ class TagFinder:
 
         return sorted(t.title for t in tags)
 
-    def genFilesWithTag(self, tagName: str | TagFilterNode,
-                        extensions: list[str] | None = None
-                        ) -> Iterable[Path]:
+    def genFilesWithTag(
+        self, tagName: str | TagFilterNode, extensions: list[str] | None = None
+    ) -> Iterable[Path]:
         if extensions is not None:
             extensions = [e if e.startswith('.') else e + '.' for e in extensions]
 
         for tagEntry in self.genEntriesWithTag(tagName):
-            if extensions is None or tagEntry.file.suffix in extensions:
-                yield tagEntry.file
+            if extensions is None or tagEntry.requireFile.suffix in extensions:
+                yield tagEntry.requireFile
 
-    def genEntriesWithTag(self, tagFilter: str | TagFilterNode) -> Iterable[TagSpaceEntry]:
+    def genEntriesWithTag(
+        self, tagFilter: str | TagFilterNode
+    ) -> Iterable[TagSpaceEntry]:
         if isinstance(tagFilter, str):
             tagFilter = TagFilterNode(tagFilter)
 
@@ -39,11 +42,15 @@ class TagFinder:
     def getAllTagEntries(self):
         return self.tss.validTagEntries
 
-    def getTagEntryInDir(self, dirPath: StrPath, recursive=True):
+    def getTagEntryInDir(self, dirPath: StrPath, *, recursive=True):
         if recursive:
-            return [te for te in self.tss.validTagEntries if te.file.is_relative_to(dirPath)]
-        else:
-            return [te for te in self.tss.validTagEntries if te.file.parent == dirPath]
+            return [
+                t
+                for t in self.tss.validTagEntries
+                if t.requireFile.is_relative_to(dirPath)
+            ]
+
+        return [t for t in self.tss.validTagEntries if t.requireFile.parent == dirPath]
 
     def getTagEntriesWithMissingConfig(self) -> Iterable[TagSpaceEntry]:
         for entries in self.tss.missingTagFiles.values():
