@@ -1,11 +1,13 @@
 from PyQt5.QtCore import QModelIndex, Qt
-from PyQt5.QtGui import QPainter, QBrush, QPalette, QColor
-from PyQt5.QtWidgets import QStyledItemDelegate, QStyleOptionViewItem, QStyle
+from PyQt5.QtGui import QBrush, QColor, QPainter, QPalette
+from PyQt5.QtWidgets import QStyle, QStyledItemDelegate, QStyleOptionViewItem
 
 
 class DarkerSelectionDelegate(QStyledItemDelegate):
-    def paint(self, painter: QPainter, option: QStyleOptionViewItem, index: QModelIndex):
-        if option.state & QStyle.State_Selected:
+    def paint(
+        self, painter: QPainter, option: QStyleOptionViewItem, index: QModelIndex
+    ):
+        if bool(option.state & QStyle.State_Selected):
             brush: QBrush = index.data(Qt.BackgroundRole)
             if brush is not None:
                 c = brush.color()
