@@ -1,13 +1,16 @@
 from enum import Enum
+from typing import TYPE_CHECKING
 
 from PyQt5.QtGui import QShowEvent
-from PyQt5.QtWidgets import QWidget, QTableWidgetItem
-
-from pyqt_settings.field.base import Field
+from PyQt5.QtWidgets import QTableWidgetItem, QWidget
 from pyqt_utils.widgets.base_ui_widget import BaseUiWidget
+
 from tag_space_tools.core.tag_finder import TagFinder
-from tag_space_tools.gui.settings import TagSpacePluginSettings, settings
+from tag_space_tools.gui.settings import TagSpacePluginSettings, tsSettings
 from tag_space_tools.ui.statistic_widget_ui import Ui_StatisticWidget
+
+if TYPE_CHECKING:
+    from pyqt_settings.field.base import Field
 
 
 class TagStatistics(Ui_StatisticWidget, BaseUiWidget, QWidget):
@@ -20,7 +23,7 @@ class TagStatistics(Ui_StatisticWidget, BaseUiWidget, QWidget):
         super().__post_init__(*args, **kwargs)
 
         field: Field = TagSpacePluginSettings.LIBRARY_PATH
-        self.libraryWidget = field.createWidgetWithLabel(settings)
+        self.libraryWidget = field.createWidgetWithLabel(tsSettings)
         self.libraryWidget.replaceWidget(self.libraryPlaceholder)
 
     def showEvent(self, showEvent: QShowEvent):
@@ -28,21 +31,23 @@ class TagStatistics(Ui_StatisticWidget, BaseUiWidget, QWidget):
         self.reloadStatistics()
 
     def reloadStatistics(self):
-        tagFinder = TagFinder(settings.LIBRARY_PATH)
+        tagFinder = TagFinder(tsSettings.LIBRARY_PATH)
         tagNameToEntries = {
             tagName: list(tagFinder.genEntriesWithTag(tagName))
-            for tagName in tagFinder.findAllTags()}
+            for tagName in tagFinder.findAllTags()
+        }
 
         self.tableWidget.clear()
         self.tableWidget.setRowCount(len(tagNameToEntries))
 
-        for i, (tagName, tagEntries) in enumerate(sorted(
-                tagNameToEntries.items(), key=lambda kv: -len(kv[1]))):
-            tagFiles = '\n'.join(str(t.file) for t in tagEntries)
+        for i, (tagName, tagEntries) in enumerate(
+            sorted(tagNameToEntries.items(), key=lambda kv: -len(kv[1]))
+        ):
+            tagFiles = '\n'.join(str(t.requireFile) for t in tagEntries)
             row = (tagName, len(tagEntries), tagFiles)
 
             column: TagStatistics.TableColumn
-            for column, val in zip(iter(self.TableColumn), row):
+            for column, val in zip(iter(self.TableColumn), row, strict=True):
                 item = QTableWidgetItem(str(val))
                 item.setToolTip(str(val))
                 self.tableWidget.setItem(i, column.value, item)
