@@ -1,33 +1,27 @@
 import json
+from collections.abc import Iterator
 from pathlib import Path
-from typing import Iterator
 
-from pyqt_utils.python.json_serializable import deepMap
 from tag_space_tools.core.tag_space_entry import Tag
 
 
 def loadTagLibrary(path: str | Path) -> Iterator[Tag]:
-    if not (path and Path(path).exists()):
+    if not path:
         return []
 
-    with open(path) as tagLibraryFile:
+    if not (libPath := Path(path)).exists():
+        return []
+
+    with libPath.open() as tagLibraryFile:
         tagLibrary = json.load(tagLibraryFile)
 
-    return parseJson(tagLibrary)
-
-
-def parseJson(tagJson) -> Iterator[Tag]:
-    if 'tagGroups' in tagJson:
-        yield from _parseTagLibraryJson(tagJson)
+    if isinstance(tagLibrary, dict) and 'tagGroups' in tagLibrary:
+        yield from _parseTagLibraryJson(tagLibrary)
     else:
-        yield from _parseTagToolsJson(tagJson)
+        yield from [Tag.fromDict(t) for t in tagLibrary]
 
 
 def _parseTagLibraryJson(tagJson: dict) -> Iterator[Tag]:
     for group in tagJson['tagGroups']:
         for tag in group['children']:
             yield Tag.fromDict(tag)
-
-
-def _parseTagToolsJson(tagJson: list) -> Iterator[Tag]:
-    yield from deepMap(tagJson, (list, Tag.fromDict))
