@@ -1,9 +1,9 @@
 from PyQt5.QtWidgets import QTabWidget
-
 from pyqt_utils.metaclass.geometry_saver import GeometrySaverMeta
 from pyqt_utils.widgets.base_ui_widget import BaseWidget
+
 from tag_space_tools.gui.duplicate.widget import DuplicateWidget
-from tag_space_tools.gui.settings import settings
+from tag_space_tools.gui.settings import tsSettings
 from tag_space_tools.gui.statistics import TagStatistics
 from tag_space_tools.gui.tag_fixer.fix_missing_tag_widget import FixMissingTagWidget
 from tag_space_tools.gui.tag_rename import TagRename
@@ -11,7 +11,7 @@ from tag_space_tools.gui.tag_sorter.widget import TagSorter
 
 
 class TagSpaceTabWidget(
-    BaseWidget, QTabWidget, settings=settings, metaclass=GeometrySaverMeta
+    QTabWidget, BaseWidget, settings=tsSettings, metaclass=GeometrySaverMeta
 ):
 
     def __post_init__(self, *args, **kwargs):
